[Bugfix] Fixed le/ge operator naming and added more operators and tests.
This fixes the `less-equal? -> less-eq?` and `greater-equal? -> greater-eq?` bug
that was causing <= and >= operators to not work.

I've additionally introduce the forward and backward relationships for
the lt/le/gt/ge for Int and Double.

I've simplified the equal/not-equal a bit.

I've introduced a unit test to at least run all the code.